Replace SurfaceHeightFacet with ElevationFacet and SurfacesFacet
