+++ src/main/java/org/terasology/flyingIslands/FlyingIslandProvider.java
@@ -6,9 +6,7 @@ package org.terasology.flyingIslands;
 import org.terasology.math.TeraMath;
 import org.terasology.math.geom.Rect2i;
 import org.terasology.math.geom.Vector2i;
-import org.terasology.math.geom.Vector2f;
 import org.terasology.utilities.procedural.Noise;
-import org.terasology.utilities.procedural.SubSampledNoise;
 import org.terasology.utilities.procedural.WhiteNoise;
 import org.terasology.world.generation.Border3D;
 import org.terasology.world.generation.Facet;
@@ -17,14 +15,14 @@ import org.terasology.world.generation.FacetProviderPlugin;
 import org.terasology.world.generation.GeneratingRegion;
 import org.terasology.world.generation.Produces;
 import org.terasology.world.generation.Requires;
+import org.terasology.world.generation.facets.ElevationFacet;
 import org.terasology.world.generation.facets.SeaLevelFacet;
-import org.terasology.world.generation.facets.SurfaceHeightFacet;
 import org.terasology.world.generation.facets.base.BaseFieldFacet2D;
 import org.terasology.world.generator.plugin.RegisterPlugin;
 
 @RegisterPlugin
 @Requires({
-        @Facet(value = SurfaceHeightFacet.class, border = @FacetBorder(sides = FlyingIsland.MAXWIDTH / 2)),
+        @Facet(value = ElevationFacet.class, border = @FacetBorder(sides = FlyingIsland.MAXWIDTH / 2)),
         @Facet(value = SeaLevelFacet.class, border = @FacetBorder(sides = FlyingIsland.MAXWIDTH / 2))
 })
 @Produces(FlyingIslandFacet.class)
@@ -36,20 +34,20 @@ public class FlyingIslandProvider implements FacetProviderPlugin {
         Border3D border = region.getBorderForFacet(FlyingIslandFacet.class).extendBy(FlyingIsland.MAXHEIGHT, FlyingIsland.MAXHEIGHT,
                 FlyingIsland.MAXWIDTH / 2);
         FlyingIslandFacet flyingIslandFacet = new FlyingIslandFacet(region.getRegion(), border);
-        SurfaceHeightFacet surfaceHeightFacet = region.getRegionFacet(SurfaceHeightFacet.class);
-        Rect2i worldRegion = surfaceHeightFacet.getWorldRegion();
+        ElevationFacet elevationFacet = region.getRegionFacet(ElevationFacet.class);
+        Rect2i worldRegion = elevationFacet.getWorldRegion();
         SeaLevelFacet seaLevelFacet = region.getRegionFacet(SeaLevelFacet.class);
 
 
         for (int wz = worldRegion.minY(); wz <= worldRegion.maxY(); wz++) {
             for (int wx = worldRegion.minX(); wx <= worldRegion.maxX(); wx++) {
-                int surfaceHeight = TeraMath.floorToInt(surfaceHeightFacet.getWorld(wx, wz));
+                int surfaceHeight = TeraMath.floorToInt(elevationFacet.getWorld(wx, wz));
                 int seaLevel = seaLevelFacet.getSeaLevel();
                 if (surfaceHeight > seaLevel && noise.noise(wx, wz) > 0.9999) {
                     FlyingIsland flyingIsland = new FlyingIsland(wx, wz);
 
                     int lowestY = surfaceHeight + 60 +  Math.floorMod(wx * wz, 30);
-                            getLowestY(surfaceHeightFacet, new Vector2i(flyingIsland.getCenter()),
+                            getLowestY(elevationFacet, new Vector2i(flyingIsland.getCenter()),
                             (int) flyingIsland.getInnerRadius(), (int) flyingIsland.getOuterRadius());
 
                     if (lowestY >= flyingIslandFacet.getWorldRegion().minY()
