--- src/main/java/org/terasology/flyingIslands/FlyingIslandTestWorldGenerator.java
@@ -1,46 +1,0 @@
-/*
- * Copyright 2020 MovingBlocks
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.terasology.flyingIslands;
-
-import org.terasology.core.world.generator.facetProviders.SeaLevelProvider;
-import org.terasology.engine.SimpleUri;
-import org.terasology.registry.In;
-import org.terasology.world.generation.BaseFacetedWorldGenerator;
-import org.terasology.world.generation.WorldBuilder;
-import org.terasology.world.generator.RegisterWorldGenerator;
-import org.terasology.world.generator.plugin.WorldGeneratorPluginLibrary;
-
-@RegisterWorldGenerator(id = "FlyingIslandsTestFlat", displayName = "FlyingIslandsTestFlat")
-public class FlyingIslandTestWorldGenerator extends BaseFacetedWorldGenerator {
-    @In
-    private WorldGeneratorPluginLibrary worldGeneratorPluginLibrary;
-
-    public FlyingIslandTestWorldGenerator(SimpleUri uri) {
-        super(uri);
-    }
-
-    @Override
-    protected WorldBuilder createWorld() {
-        return new WorldBuilder(worldGeneratorPluginLibrary)
-                .addProvider(new SurfaceProvider())
-                .addProvider(new SeaLevelProvider(0))
-                .addProvider(new FlyingIslandProvider())
-                .addRasterizer(new SurfaceRasterizer())
-                .addRasterizer(new FlyingIslandRasterizer())
-                .addPlugins();
-    }
-
-}
